refactor(driver)!: move driver registration to sub-packages
We remove the init-function which registers the SQL driver when
importing the main package `github.com/golistic/pxmysql`. Now, users
must register (if wanted), using `github.com/golistic/pxmysql/register`.

If the "mysql" driver name needs to be registered, users can anonymously
import `github.com/golistic/pxmysql/register/mysql`.

BREAKING CHANGING:
With moving the registration to sub-packages, users will need to
import (anonymously) `github.com/golistic/pxmysql/register`.
We do not keep backwards compatibility.

Resolves #50
