+++ _support/pxmysql-compose/shared/goapps/unix_socket/main.go
@@ -17,7 +17,7 @@ import (
 	"fmt"
 	"log"
 
-	_ "github.com/golistic/pxmysql"
+	_ "github.com/golistic/pxmysql/register"
 )
 
 func main() {
