+++ register/register_test.go
@@ -1,6 +1,6 @@
 // Copyright (c) 2023, Geert JM Vanderkelen
 
-package mysql
+package register
 
 import (
 	"database/sql"
@@ -15,7 +15,7 @@ func TestDriver_Open(t *testing.T) {
 		xt.Assert(t, xstrings.SliceHas(sql.Drivers(), "pxmysql"), "expected driver pxmysql to be registered")
 	})
 
-	t.Run("mysql is registered", func(t *testing.T) {
-		xt.Assert(t, xstrings.SliceHas(sql.Drivers(), "mysql"), "expected driver mysql to be registered")
+	t.Run("mysql is not registered", func(t *testing.T) {
+		xt.Assert(t, !xstrings.SliceHas(sql.Drivers(), "mysql"), "expected driver mysql not to be registered")
 	})
 }
